requirements.txt
Flask==2.2.3
Werkzeug==2.3.7
pymongo==4.5.0
python-dotenv==1.0.0
flask-mongoengine==1.0.0
dnspython==2.4.2
Flask-WTF==1.1.1
